+++ CHANGELOG.md
@@ -17,6 +17,14 @@ and this project adheres to
 
 ### Added
 
+### Changed
+
+### Fixed
+
+## [v2.1.0] - 2024-03-20
+
+### Added
+
 - TSVector index to log_lines, and gin index to dataclips
   [#1898](https://github.com/OpenFn/lightning/issues/1898)
 - Add API Version field to Salesforce OAuth credentials
@@ -26,8 +34,6 @@ and this project adheres to
 
 - Replace v1 usage tracking with v2 usage tracking.
   [#1853](https://github.com/OpenFn/lightning/issues/1853)
-
-### Fixed
 
 ## [v2.0.10]